Sending the 1st draft of the Exxos / Kernel layers to the main branch so it can be used right after for individual tasks.
The actual Exxos layer is registry-based: it uses a "components registry" to know where the classes are located in the filesystem. This way of doing has a few little inconvenients, but I think they are acceptable.
IMO this does work well enough to be used, so now real "features" can be done... I'll be keeping the more "revolutionnary" patterns that only harcore coders can really appreciate for a next version.

Special note about this commit: old-style error handling has been disabled... something will come back later (but it's a work branch, so no problem).