Added a few helper methods in the smarty compiler, for use by compiler plug-ins.
I've refactored many classes (i.e: all the form elements) to compiler extensions, to prevent having to load these classes during runtime (this way the element rendering code is inserted inside the compiled template, and there is no need to load the class definition anymore when the template is displayed). Now, due to the way compiler extensions get their parameters, this was not particularly easy or looking nice, and the methods added here are meant to help a little.

Nothing is documented, as this is still very unstable... at least it works, but it is still a bit too ugly to be something I can consider definitive for the moment.