Added a new mechanism to load Smarty plug-ins.
The problem: the Smarty plug-in loading mechanism is not flexible at all... all plug-ins have to be copied to a particular folder to be found, and the list of scanned folders are specified on the PHP side (so it's impossible to add 3rd party plug-ins without copying them to one of these folders, or hack something).

The additional mechanism allows to override this and to dynamically load a set of plug-ins by just adding the correct instruction to templates. It works that way:
- You can load a "smarty extension" by using the ([xoLoadExtension your_class_name]) instruction somewhere on top of your template
- This class gets instanciated, and its extendTemplate() method is called
- The extendTemplate() method then registers its plug-ins and makes them available to the tpl engine/compiler

This duplicates a little some of the provided Smarty functionnality, but really brings some welcome flexibility and allows things that would not have been able by just making some kind of ([xoAddPluginsFolder /path/to/custom/plugins]) plug-in... also, as most of the things I'm doing to Smarty, the real logic is executed during the compilation phase and has no serious consequences on the runtime scalability/speed.