Adding the pyro widgets library template extension.
When loaded in templates using ([xoLoadExtension xoops_pyro]), it ensures the prototype and scriptaculous libraries get loaded, and registers all the form fields widgets to Smarty.
It also provides 2 additional features:
- An ([xoForm]) Smarty plug-in that allows form fields to know which form model they are referring to...
Must be used like this:
([xoForm $formTplVar])
  ([xoInput ref='login'])
  ([xoSelect1 ref='rememberMe'])
([/xoForm])

- A behavior-like mechanism that allows widgets using Javascript to work without cluttering the document mark-up with <script> crap.
On document load, HTML elements are checked for a class named "pyro_Something", and if a Javascript class with the same name is found, an instance of it is created for the said element.

So, is the document contains:
<div id="mytree" class="pyro_TreeWidget">....</div>

Then something like this will be automatically done during onload:
new pyro_TreeWidget( document.getElementById( "mytree" ) );
Without having to  manually add this snippet in a script tag.

PS: The choice to go for prototype is far from definitive... there are some great features there, but IMO a lot of useless stuff just duplicating already avail JS funcs, so I'm not sure about that.
It's not used in every page and only loaded for advanced uses, and is becoming some kind of standard library in the ajax-web50000-buzzwords-whatever-Ithink-Im-smart-coz-I-have-a-blog world, but I still wonder a lot...