Added the HTTP handler schemas(s).
The XSD has been written manually, the XML version is generated from it using an XSL that transformed it to something more canonical, and then serialized to a PHP array structure that can be used during runtime.
The validator class of the HTTP config panel has been generated from this schema, but more generators are expected to come soon.

I'll commit the transformation / generation tools later, as I'm still experimenting with that right now.