Added a front controller like class to dispatch request processing.
There are two classes provided: one base class (xos_app_RequestDispatcher) providing the interface, and a generic implementation (xos_app_VarRequestDispatcher) that uses the value of a specific request variables to choose which page/action controller to instanciate.

The later has been made for the session panel (which is made of two pages), but has been thought generically enough to be re-used later on when dealing with content mgmt stuff (to dispatch requests to action controllers).