"Created a new ModuleInstaller class
Am going to move most of the modules installation panel to this class.

At least this precise functionality has to be taken care of right now,
as the modulatisation of some core features will require the installation of some modules
from the install wizard (and it will be needed immediately, as the work done on this branch
will give birth to a \"workbench\" module that needs to be installed).

This class is supposed to make this easy."