Removed XoopsSimpleForm and XoopsThemeForm
Here's the point: I needed to make a custom XoopsForm class, and the actual system forced me to make it a XoopsThemeForm child which I didn't like at all.

So I just added a new "viewMode" property to XoopsForm and made its render() method concrete by moving the XoopsSimpleForm/XoopsThemeForm code to it. This way of doing it is not 100% clean but as the "xoopsformloader" stuff is real real ugly it doesn't matter at all and the result not worse than what we had before actually (well, normally I'd put a smiley here, but I don't know if it should be a happy or real sad one actually :S).

The old classes have been kept for BC purpose but are empty, so people can freely instanciate XoopsForm directly now.