Added a new XoopsConfigForm class to edit XoopsConfig options
I just moved the code that was in system/admin/preferences/main.php to a dedicated class to clean up the admin panel code.

This form can automatically create the elements needed to edit the config options you send to its addElements() method. It's normally only meant to be used by the old preferences panel, but could theorically be reused by 3rd party modules that would find it usefull too.