Fixed bug #1517269: Bad login redirection when XOOPS is inside a subfolder
Small comment about this one: the reported bug was not entirely a bug, but more the an indication of a wrong behavior.
The problem was this:
xoops_redirect had to be a semi-absolute URL (so, for a XOOPS installed in domain.tld/subfolder xoops_redirect had to be /subfolder/targetpage.php) while we would expect /targetpage.php to work.

So, I fixed the redirection code, to accept "good" URLs while still being able to deal with the old case (it checks the xoops_redirect value: if it starts with "/subfolder" it's expected to be an old-school semi-abs URL).