Added global WYSIWYG extensions support
This patch allows XoopsFormDhtmlTextArea to delegate rendering to an external editor.

It's disabled by default and thus shouldn't bring any particular issue.
To enable this system, one line has to be changed manually here (as there's no config UI provided).

I'll mark this as experimental in the release notes but will explain how to use it, so volunteers can test it and we can see if it's worth using later.