Refactored the last installer steps
Instead of one "tables creation" and one "data insertion" step, the last pages of the wizard are now "setup system" and "setup modules".

The system is just made of low-level stuff (the users, session, modules, etc... tables as well as admin user creation), and everything else is performed in the modules setup.