Fixing DJ's XoopsMultiMailer::SetLanguage implementation
2nd part: just changing SetLanguage like that is a bit useless... it would still have to be called correctly and this isn't done anywhere. 

So this 2nd change tries to ensure the PHPMailer language is automatically set to the correct one if possible: XoopsMultiMailer will call SetLanguage() automatically when anew instance is created, and the _LANGCODE value will be used there.

There are just 2 problems with that:
- If I'm correct, the brasilian XOOPS version has a langcode of pt_br and not "br" as used in the phpmailer langfiles.
- There are 2 chinese files, with the encoding added to the name, so just using _LANGCODE wont work there either

I think it would be OK to add another "if" to handle those 2 cases specifically, but I dont have a brasilian or chinese XOOPS available right now, so I cant do it. Anyway, if somebody wants to do it on Tuesday that would be OK I guess ( I announced 2.0.16 for Wednesday, and I think I wont package it before the Wed/Thirs night ).