Added the PHPMailer language files that were not sent to the trunk.
REMINDER: The current dev version is the trunk. So every single bugfix sent on a maintenance branch should also be sent here, unless there's a specific reason not to.

So, I guess it's OK not to merge the recent session modif here as session management has been rewritten in the 2.3 alpha branch, but this mailer stuff should have been sent to the trunk when sent to the 2016 branch.