Modified XoopsMultiMailer so it looks for translated versions of its messages in /language/currentlang/phpmailer.php
XOOPS translators can now include a custom version of the PHPMailer strings in their language folder, and it will be used if found.

This file format is a bit different from the regular XOOPS strings files, but it should be easy to handle.
Apart from this, it should be written like all other XOOPS translation files: so make sure it uses the same encoding as the rest of your language pack files.

PS: The default (english) version is located in /class/mail/phpmailer/language/phpmailer.lang-en.php and can be used as a template to make translations.
PS2: If you're lazy, you could use one the files files distributed with the PHPMailer class... just make sure the file you provide is encoded like the rest of your language pack.