Created the md5 checksum creation script.
This is another CLI script that calculates md5 checksums of the XOOPS core files.

The routine is still the one used in the original script, it has just been slightly changed to take the root folder name into account.
i.e: before "mainfile.php" was just called "mainfile.php" in the md5 file, and now it is called "htdocs/mainfile.php"

This is necessary for future versions that will contain more than 1 folder to check.