Deleting md5 script / checksum file from work branches...
Having those here is pointless, and actually makes it harder to track changes to the script.

So I'm moving the script to /XoopsCore/ along with the other svn tools, and it will just be added to the XoopsCore/releases/X.Y.Z/ folder by the releases creation script each time a release is packaged.