First version of the releases creation script
The script should now be over... I enhanced it a little so there's now a "simulation mode" where commands will be displayed but not executed, a "comfirm mode" where you'll be asked for confirmation before each command is actually run.

It's also menu-driven, so the different phases ( tagging / md5 generation / packages creation ) can be performed separately (can be useful if you need to perform something special before packaging the files).