Eliminate PHP 4 style constructor
Several classes had code similar to this:

```
class XoopsDatabaseFactory
{

    public function __construct()
    {
    }

    public function XoopsDatabaseFactory()
    {
        $this->__construct();
    }
```

In this example, In PHP4 (which we do not support,) `function XoopsDatabaseFactory()` would the constructor. Now it serves no purpose, as it cannot be invoked without having already instantiated the class (so `__construct()` was already invoked.)

Also, some visibility changes where appropriate.