Replace getAll() in ProfileVisibilityHandler
Since the profile_visibility table does not have a single unique
column, the *id_as_key* logic will not work correctly. Added a
getAllByFieldId() method to return an array of arrays of rows,
which is what the template was expecting.

This change gives us a complete set of visibilities for each field.