Make permission op sticky
The admin/permission tab resets to the edit op after every
update. This takes away the opportunity for the admin to verify
any changed settings post submit without additional action.

This change makes the "op" sticky, so that after submitting
changes, the page will display the data that was just changed.