Add a table property to handlers
If these were implemented as XoopsPersitableObjectHandlers, as
they should have been, there would be a public table property.
Certain programs access 'core' tables by name rather than
through handlers. This causes issues when those tables are
renamed (as they are in 2.6.) Using $handler->table to access
the table name is possible in 2.6, this makes it possible in 2.5.