Add cookie domain to install
The domain of a cookie effects which subdomains the cookie will
be sent to.

Historically, the XOOPS_COOKIE_DOMAIN constant has been
determined by an algorithm, which in some cases produced
invalid results. Changing it required digging into code.
The constant will now be explicitly defined mainfile.php
and exposed as a configurable option during the install.