Add rendered form to XoopsForm::assign()
Add a fully rendered version of the form to the template
assignments when the XoopsForm::assign() method is used.
The form rendering is added to the 'rendered' key of the
assignments.

This change allows the theme templates to choose the full system
rendered form or process the individual elements as appropriate.