Remove first time redirect
Remove the admin side redirect that triggers anytime the
'transition_theme' cookie does not exist.

The upgrade process "deep links" to the module update for the
system module. It is possible that other code in the admin area
will not work correctly until the update is performed, and the
redirect keeps the admin user from completing the upgrade,
potentially leaving a non-functional system.