Sanitize Mime-Type
In XoopsMediaUploader, there is an option, $allowUnknownTypes,
that could allow a malicious mime-type supplied in an HTTP
request to be returned directly to the caller. Without further
processing by the caller, a carefully crafted mime-type could
be used to facilitate other malicious actions.

This change checks the structure of the mime-type, and will
always reject any upload attempted with a non-conforming
content-type specified (i.e. containing script code.)

Fixes #524