Move protector data to xoops_data
Protector requires write access to xoops_lib/modules/protector.
XOOPS_VAR_PATH is always intended to be writable so that it can
be used for files like those that protector uses.

This changes protector's configs to xoops_data/protector.

Note: existing protector configs are not moved. The file names
remain the same, so if needed (i.e. IP ban lists) the files can
be moved manually.