Do not load jquery.js if already loaded
This change makes the loading of jquery for jgrowl support
conditional on jQuery not being defined.

Previously, when jgrowl was loaded it replaced any previously
loaded jquery instance with a new one. This resulted in some
interference with other code, such as bootstrap, that had been
loaded earlier in the page.