Allow comment instead of blank title
Suppress display of block titles starting with `// `.

The display of a block title can be suppressed by leaving the
title field blank, but this makes it difficult to identify
blocks when editing in the admin section.

This change allows "commenting" out a title by starting it with
a slash-slash-space sequence. Blocks with a title starting with
this sequence will be displayed as if their title was blank.