Remove locale specific date format for calendar.js
Several issues found:

The localized format date is directly used to instantiate
a js Date object, and those localized formats are not
supported in that context.

There seems to be a bug that effects the preferred ISO format,
causing calendar.js to be off by one day back(?)

The generally supported mm/dd/yyyy seems to work correctly.

https://www.w3schools.com/js/js_date_formats.asp