Add XoopsHttpGet
This class is used to retrieve the response from an http(s)
request to a URL. It is replacing "Snoopy" that has been used
for fetching rss feeds in the administration area.

For proper operation, it requires the curl PHP extension, or
alternately that the ini settiing 'allow_url_fopen' is enabled.