Recognize error return from db->query()
\XoopsModelRead::getAll() fails to handle a return of false
from a database query(). It then attempts to use the false
value as a result set which now causes a nasty failure.

As there is no documented failure return mechanism, the
return in the case of failure is now an empty array. This
may not be ideal, but should not break compatibility.

Also, removed some dead code regarding an $orderSet variable.