Errors using XOOPS_VAR_PATH early in install
Before mainfile.php is written, XOOPS_ROOT_PATH is defined
based on the filesystem location of installer. The forms used
to collect install data use textsanitizer in some situations.
Textsanitizer now uses XOOPS_VAR_PATH in locating config files,
and in this case, before that define is established.

The only method used is MyTextSanitizer::htmlSpecialChars
so this commit replaces those calls with calls to a work-alike
local function.