Update theme.tpl
In light/dark mode, there was a problem with overloads if the definitions were the same in my_xoops.css and my_xoops_dark.css. The system was reading the last declared definitions, so in my_xoops_dark.css, even if the system was in light mode.
Now, there is no more overwriting between the 2 files