Fix non-conforming id elements in module menu tabs
Was mentioned on in #1133 but not properly dealt with

The menu tab id was built from the display text, which varies
with language in use. Also no attempt to conform to HTML spec
leaving spaces and other non-permitted characters.

Id values now start with 'modmenu_' and then an integer based
on position 0-n. The active menu tab still has id attribute
of 'current'