Fix issues with Bad IP list in Protector Center
Replaces sort function with one capable of handling both IPv4
and IPv6 addresses.

The inclusion of IPv6 addresses required a small change in how
short duration ban times are shown in the Bad IPs list. The
separator between the address and time was a ':' and is now a
'|'. IPv6 addresses use the colon, so this needed to be changed.

Also some small UI tweaks to make display of IPv6 addresses more
readable.

Add Chrome to the short useragent logic

Fixes #1151