update siteinfo bloc and 1 question
The test  of $block.showgroups should be before the table tag.
Why ?
Cause if $block.showgroups == false, the result is a useless table in the code.

I have a question on another point :
Why test "        <!-- start group loop -->
        <{foreach item=group from=$block.groups}>
" ?
Cause there is only the webmasters group display, no other group in this bloc.
So, I think this test is useless, I think we should pick yp the group webmasters, no need to use a foreach for that.
No ?